Move triage and PSD instruction to PatientProgramme
