+++ app/models/instruction.js
@@ -1,6 +1,6 @@
 import { fakerEN_GB as faker } from '@faker-js/faker'
 
-import { PatientSession, Programme } from '../models.js'
+import { Patient, Programme } from '../models.js'
 
 import { BaseModel } from './base.js'
 
@@ -26,10 +26,10 @@ export class Instruction extends BaseModel {
     super(options, context)
 
     /** @type {string|undefined} */
-    this.patientSession_uuid
+    this.patient_uuid
 
-    /** @type {PatientSession|undefined} */
-    this.patientSession
+    /** @type {Patient|undefined} */
+    this.patient
 
     /** @type {string|undefined} */
     this.programme_id
@@ -42,11 +42,7 @@ export class Instruction extends BaseModel {
   }
 }
 
-Instruction.relate(
-  'patientSession_uuid',
-  () => PatientSession,
-  'patientSession'
-)
+Instruction.relate('patient_uuid', () => Patient, 'patient')
 Instruction.relate('programme_id', () => Programme, 'programme')
 
 /**
