+++ app/models/patient-session.js
@@ -28,7 +28,6 @@ import {
   ClinicAppointment,
   ClinicBooking,
   Gillick,
-  Instruction,
   Patient,
   Programme,
   Session
@@ -104,12 +103,6 @@ export class PatientSession extends BaseModel {
    */
   constructor(options, context) {
     super(options, context)
-
-    /** @type {string|undefined} */
-    this.instruction_uuid
-
-    /** @type {Patient|undefined} */
-    this.instruction
 
     /** @type {string|undefined} */
     this.patient_uuid
@@ -878,69 +871,6 @@ export class PatientSession extends BaseModel {
   }
 
   /**
-   * Record triage
-   *
-   * @param {Partial<AuditEvent>} event - Event
-   */
-  recordTriage(event) {
-    this.patient?.addEvent({
-      name: activity.triage.decision(event),
-      note: event.note,
-      type: AuditEventType.ProgrammeNote,
-      outcome: event.outcome,
-      outcomeAt_: event.outcomeAt_,
-      createdAt: event.createdAt,
-      createdBy_uid: event.createdBy_uid,
-      programme_ids: [this.programme_id]
-    })
-
-    let messageTemplate
-    switch (event.outcome) {
-      case ScreenOutcome.DelayVaccination:
-        messageTemplate = 'triage-delay-vaccination'
-        break
-      case ScreenOutcome.DoNotVaccinate:
-        messageTemplate = 'triage-do-not-vaccinate'
-        break
-      case ScreenOutcome.InviteToClinic:
-        messageTemplate = 'triage-invite-to-clinic'
-        break
-      default:
-        messageTemplate = 'triage-vaccinate'
-    }
-
-    if (this.patient?.contacts) {
-      for (const contact of this.patient.contacts) {
-        this.patient?.addEvent({
-          name: activity.notify[messageTemplate](contact),
-          messageRecipient: contact,
-          messageTemplate,
-          createdAt: event.createdAt,
-          patient_uuid: this.uuid,
-          programme_ids: [this.programme_id],
-          session_id: this.session?.id
-        })
-      }
-    }
-  }
-
-  /**
-   * Give PSD instruction
-   *
-   * @param {Instruction} instruction - Instruction
-   */
-  giveInstruction(instruction) {
-    this.instruction_uuid = instruction.uuid
-
-    this.patient?.addEvent({
-      name: activity.psd.added,
-      createdAt: instruction.createdAt,
-      createdBy_uid: instruction.createdBy_uid,
-      programme_ids: [this.programme_id]
-    })
-  }
-
-  /**
    * Register attendance
    *
    * @param {Partial<AuditEvent>} event - Event
@@ -1011,7 +941,6 @@ export class PatientSession extends BaseModel {
   }
 }
 
-PatientSession.relate('instruction_uuid', () => Instruction, 'instruction')
 PatientSession.relate('patient_uuid', () => Patient, 'patient')
 PatientSession.relate('programme_id', () => Programme, 'programme')
 PatientSession.relate('session_id', () => Session, 'session')
