+++ app/routes/patient-session.js
@@ -17,7 +17,6 @@ router.post('/:nhsn/:programme_id/new/gillick', patientSession.gillick('new'))
 router.post('/:nhsn/:programme_id/new/pre-screen', patientSession.preScreen)
 router.post('/:nhsn/:programme_id/new/invite', patientSession.invite)
 router.post('/:nhsn/:programme_id/new/remind', patientSession.remind)
-router.post('/:nhsn/:programme_id/new/triage', patientSession.triage)
 router.post('/:nhsn/:programme_id/new/note', patientSession.note)
 
 router.all('/:nhsn/:programme_id/edit/:view', patientSession.readForm)
@@ -25,7 +24,6 @@ router.get('/:nhsn/:programme_id/edit/:view', patientSession.showForm('edit'))
 
 router.post('/:nhsn/:programme_id/edit/gillick', patientSession.gillick('edit'))
 router.post('/:nhsn/:programme_id/edit/registration', patientSession.register)
-router.post('/:nhsn/:programme_id/edit/triage', patientSession.triage)
 
 router.get('/:nhsn/:programme_id{/:view}', patientSession.show)
