+++ lib/create-data.js
@@ -432,7 +432,7 @@ for (const patientSession of Object.values(context.patientSessions)) {
         }
 
         // 4️⃣ SCREEN with triage outcome (initial)
-        patientSession.recordTriage({
+        patientSession.patientProgramme.recordTriage({
           outcome,
           note,
           createdAt: response.createdAt,
@@ -460,7 +460,7 @@ for (const patientSession of Object.values(context.patientSessions)) {
       context.instructions[instruction.uuid] = instruction
 
       // GIVE INSTRUCTION for PSD
-      patientSession.giveInstruction(instruction)
+      patientSession.patientProgramme.giveInstruction(instruction)
     }
   }
 
@@ -469,7 +469,7 @@ for (const patientSession of Object.values(context.patientSessions)) {
     // Ensure any outstanding triage has been completed
     if (patientSession.screen === ScreenOutcome.NeedsTriage) {
       // 4️⃣ SCREEN with triage outcome (final)
-      patientSession.recordTriage({
+      patientSession.patientProgramme.recordTriage({
         outcome: ScreenOutcome.Vaccinate,
         note: 'Spoke to GP, safe to vaccinate.',
         createdAt: removeDays(session.date, 2),
